+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed styles and labels for navigation buttons.
 - Added lesson title into section above lesson content.
 - Improved styles for disabled buttons.
+- Removed toggleable "Hide modules" / "Show modules" course navigation link. Now the navigation is static.
+- Removed custom hook useLocalStorage() since it's not used anywhere in the code anymore
+- Changed css "position" of the course navigation from absolute to static since it does not have to be absolute anymore.
+- Changed styles of an active lesson / quiz item in the course navigation
 
 ## [2.6.4]
 - Added automatic switching to the section with search keywords if the first section does not contain any